begin outline test code for the main work of `cpr`
