+++ pending_review/review_test.go
@@ -188,3 +188,106 @@ func TestKnowCase6144(t *testing.T) {
 		},
 	}, result)
 }
+
+func TestKnowCase16144(t *testing.T) {
+	reviews := parseReviewJSON(t, `[
+		{
+		  "id": 1311404502,
+		  "node_id": "PRR_kwDODDMJAM5OKnHW",
+		  "user": {
+			"login": "toge",
+			"id": 465629,
+			"node_id": "MDQ6VXNlcjQ2NTYyOQ==",
+			"avatar_url": "https://avatars.githubusercontent.com/u/465629?u=fc95d16a396044be5625091463f1f89711bdc05e&v=4",
+			"gravatar_id": "",
+			"url": "https://api.github.com/users/toge",
+			"html_url": "https://github.com/toge",
+			"followers_url": "https://api.github.com/users/toge/followers",
+			"following_url": "https://api.github.com/users/toge/following{/other_user}",
+			"gists_url": "https://api.github.com/users/toge/gists{/gist_id}",
+			"starred_url": "https://api.github.com/users/toge/starred{/owner}{/repo}",
+			"subscriptions_url": "https://api.github.com/users/toge/subscriptions",
+			"organizations_url": "https://api.github.com/users/toge/orgs",
+			"repos_url": "https://api.github.com/users/toge/repos",
+			"events_url": "https://api.github.com/users/toge/events{/privacy}",
+			"received_events_url": "https://api.github.com/users/toge/received_events",
+			"type": "User",
+			"site_admin": false
+		  },
+		  "body": "",
+		  "state": "APPROVED",
+		  "html_url": "https://github.com/conan-io/conan-center-index/pull/16144#pullrequestreview-1311404502",
+		  "pull_request_url": "https://api.github.com/repos/conan-io/conan-center-index/pulls/16144",
+		  "author_association": "CONTRIBUTOR",
+		  "_links": {
+			"html": {
+			  "href": "https://github.com/conan-io/conan-center-index/pull/16144#pullrequestreview-1311404502"
+			},
+			"pull_request": {
+			  "href": "https://api.github.com/repos/conan-io/conan-center-index/pulls/16144"
+			}
+		  },
+		  "submitted_at": "2023-02-23T14:15:27Z",
+		  "commit_id": "e2aa65c961d48d688dd5450811229eb1d62649ba"
+		},
+		{
+		  "id": 1335829632,
+		  "node_id": "PRR_kwDODDMJAM5PnySA",
+		  "user": {
+			"login": "prince-chrismc",
+			"id": 16867443,
+			"node_id": "MDQ6VXNlcjE2ODY3NDQz",
+			"avatar_url": "https://avatars.githubusercontent.com/u/16867443?u=d7c5b45b864fe5d26e44d38645a22f8af18a2a16&v=4",
+			"gravatar_id": "",
+			"url": "https://api.github.com/users/prince-chrismc",
+			"html_url": "https://github.com/prince-chrismc",
+			"followers_url": "https://api.github.com/users/prince-chrismc/followers",
+			"following_url": "https://api.github.com/users/prince-chrismc/following{/other_user}",
+			"gists_url": "https://api.github.com/users/prince-chrismc/gists{/gist_id}",
+			"starred_url": "https://api.github.com/users/prince-chrismc/starred{/owner}{/repo}",
+			"subscriptions_url": "https://api.github.com/users/prince-chrismc/subscriptions",
+			"organizations_url": "https://api.github.com/users/prince-chrismc/orgs",
+			"repos_url": "https://api.github.com/users/prince-chrismc/repos",
+			"events_url": "https://api.github.com/users/prince-chrismc/events{/privacy}",
+			"received_events_url": "https://api.github.com/users/prince-chrismc/received_events",
+			"type": "User",
+			"site_admin": false
+		  },
+		  "body": "",
+		  "state": "APPROVED",
+		  "html_url": "https://github.com/conan-io/conan-center-index/pull/16144#pullrequestreview-1335829632",
+		  "pull_request_url": "https://api.github.com/repos/conan-io/conan-center-index/pulls/16144",
+		  "author_association": "CONTRIBUTOR",
+		  "_links": {
+			"html": {
+			  "href": "https://github.com/conan-io/conan-center-index/pull/16144#pullrequestreview-1335829632"
+			},
+			"pull_request": {
+			  "href": "https://api.github.com/repos/conan-io/conan-center-index/pulls/16144"
+			}
+		  },
+		  "submitted_at": "2023-03-11T06:46:57Z",
+		  "commit_id": "e2aa65c961d48d688dd5450811229eb1d62649ba"
+		}
+	  ]`)
+	reviewers := ConanCenterReviewers{Reviewers: []Reviewer{
+		{User: "czoido", Type: "team", Requested: false},
+		{User: "memsharded", Type: "team", Requested: false},
+		{User: "uilianries", Type: "team", Requested: true},
+		{User: "SpaceIm", Type: "community", Requested: false},
+		{User: "ericLemanissier", Type: "community", Requested: false},
+		{User: "prince-chrismc", Type: "team", Requested: false},
+		{User: "Croydon", Type: "community", Requested: false},
+		{User: "toge", Type: "community", Requested: false},
+	}}
+	result := ProcessReviewComments(&reviewers, reviews, "e2aa65c961d48d688dd5450811229eb1d62649ba")
+	assert.Equal(t, Reviews{
+		Count: 2, ValidApprovals: 2, TeamApproval: true,
+		Approvals: []string{"toge", "prince-chrismc"},
+		Blockers:  nil, LastReview: &Review{
+			ReviewerName: reviews[len(reviews)-1].GetUser().GetLogin(),
+			SubmittedAt:  reviews[len(reviews)-1].GetSubmittedAt(),
+			HTMLURL:      reviews[len(reviews)-1].GetHTMLURL(),
+		},
+	}, result)
+}
